8-8-2023 changes.
Implemented the delete playlist function. By doing this I made it so that the user can delete the head playlist node which crashes the program if you try and print the playlist.

My current solution is to set the name of it to be an empty string but I guess that also crashes it. My solution to this will be to rename it to just a space. Not sure.