8-9-2023 current changes
My solution to the user being able to delete the head node and destroy the program was to make a head node that the user does not know about. The user cannot delete this playlist or manipulate it unless they know that the playlist name is "HEAD92903449". 

The reason behind the numbers is that it is just a random prime number I used from my prime number generator.

As of now, this has no flaws.